% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodyweight.R
\docType{data}
\name{bodyweight}
\alias{bodyweight}
\title{Trial Mice Bodyweight}
\format{
A tibble with 32 rows and 7 variables:
\describe{
\item{ID}{a factor denoting penguin species (Adélie, Chinstrap and Gentoo)}
\item{Body.Weight.1}{a factor denoting island in Palmer Archipelago, Antarctica (Biscoe, Dream or Torgersen)}
\item{Date.Body.Weight.1}{a number denoting bill length (millimeters)}
\item{Body.Weight.2}{a number denoting bill depth (millimeters)}
\item{Date.Body.Weight.2}{an integer denoting flipper length (millimeters)}
\item{Body.Weight.3}{an integer denoting body mass (grams)}
\item{Date.Body.Weight.3}{a factor denoting penguin sex (female, male)}
}
}
\source{
{Adélie penguins: Palmer Station Antarctica LTER and K. Gorman. 2020. Structural size measurements and isotopic signatures of foraging among adult male and female Adélie penguins (Pygoscelis adeliae) nesting along the Palmer Archipelago near Palmer Station, 2007-2009 ver 5. Environmental Data Initiative.} \doi{10.6073/pasta/98b16d7d563f265cb52372c8ca99e60f}

{Gentoo penguins: Palmer Station Antarctica LTER and K. Gorman. 2020. Structural size measurements and isotopic signatures of foraging among adult male and female Gentoo penguin (Pygoscelis papua) nesting along the Palmer Archipelago near Palmer Station, 2007-2009 ver 5. Environmental Data Initiative.} \doi{10.6073/pasta/7fca67fb28d56ee2ffa3d9370ebda689}

{Chinstrap penguins: Palmer Station Antarctica LTER and K. Gorman. 2020. Structural size measurements and isotopic signatures of foraging among adult male and female Chinstrap penguin (Pygoscelis antarcticus) nesting along the Palmer Archipelago near Palmer Station, 2007-2009 ver 6. Environmental Data Initiative.} \doi{10.6073/pasta/c14dfcfada8ea13a17536e73eb6fbe9e}

{Originally published in: Gorman KB, Williams TD, Fraser WR (2014) Ecological Sexual Dimorphism and Environmental Variability within a Community of Antarctic Penguins (Genus Pygoscelis). PLoS ONE 9(3): e90081. doi:10.1371/journal.pone.0090081}
}
\usage{
bodyweight
}
\description{
This dataset documents the bodyweight of our trial mice
}
\keyword{datasets}
