% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footballbetting23.R
\docType{data}
\name{footballbetting23}
\alias{footballbetting23}
\title{Football '23 Season Betting}
\format{
A tibble with 285 rows and 18 variables:
\describe{
\item{schedule_date}{A date string denoting the date that the NFL game was played}
\item{schedule_season}{A numeric value denoting the year of the NFL season}
\item{schedule_week}{A string denoting the week of the game in the regular season or the post season round}
\item{schedule_playoff}{A logical value denoting if the game was a playoff game}
\item{team_home}{A string denoting the home NFL team for the game}
\item{score_home}{An integer denoting the points scored by the home team}
\item{score_away}{An integer denoting the points scored by the away team}
\item{team_away}{A string denoting the away NFL team for the game}
\item{team_favorite_id}{A string denoting the ID (abbreviation) of the team favored in the game by sportsbooks}
\item{spread_favorite}{A numeric value denoting how many points the favorite team was favored by sportsbooks}
\item{over_under_line}{A numeric value denoting the projection for total points expected by sportsbooks}
\item{stadium}{A string denoting the stadium where the game was played}
\item{stadium_neutral}{A logical value denoting if the game was hosted at a neutral site}
\item{weather_temperature}{A numeric value denoting the temperature during the game}
\item{weather_wind_mph}{A numeric value denoting the wind speed during the game}
\item{weather_humidity}{A numeric value denoting the humidity of the stadium during the game}
\item{weather_detail}{A string denoting where the game was played (indoor v. outdoor)}
\item{over_under}{A string denoting if the game resulted in the over/under/push compared to the expected over/under line set by sportsbooks}
}
}
\source{
{Dataset sourced from Kaggle: https://www.kaggle.com/datasets/tobycrabtree/nfl-scores-and-betting-data?resource=download&select=spreadspoke_scores.csv}
}
\usage{
footballbetting23
}
\description{
This dataset documents outcomes and betting information for all NFL games of the '23 season.
}
\keyword{datasets}
