% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_columns.R
\name{summarize_columns}
\alias{summarize_columns}
\title{Summarizes the Class and Missingness of Columns in a Dataframe}
\usage{
summarize_columns(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A summary table containing each column name along with its class and number of missing values.
}
\description{
\code{summarize_columns()} finds the class of all columns in a dataframe and the number of missing values
}
\examples{
summarize_columns(footballbetting23)
summarize_columns(bodyweight)

}
