% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_outlier_mice.R
\name{track_outlier_mice}
\alias{track_outlier_mice}
\title{Title}
\usage{
track_outlier_mice(
  data,
  id_col = "ID",
  weight_cols = c("Body.Weight.1", "Body.Weight.2", "Body.Weight.3"),
  threshold = 0.2
)
}
\arguments{
\item{threshold}{}
}
\description{
Title
}
