% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_betting_metrics.R
\name{validate_betting_metrics}
\alias{validate_betting_metrics}
\title{Title}
\usage{
validate_betting_metrics(data, metric_cols, ranges = list())
}
\arguments{
\item{ranges}{}
}
\description{
Title
}
